ALRP/requirements.txt
numpy
pandas
flask
typing
imutils
pytesseract
pillow
matplotlib